+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * add `define-record-type`
 * add `escape-regex` function
 * make `apropos` accept symbol as argument
+* add doc strings for `**interaction-environment**` and `**internal-env**`
 ### Bugfix
 * fix prism highlighting of names (for new context help)
 * fix error when using help in node REPL and there are no doc string